+++ scripts/libs/fw_ver_2_2_x.py
@@ -134,10 +134,7 @@ class register_mem_cfg_t(Structure):
                 ("internal_vref11", c_internal_vref11),
                 ("internal_temp_calibration", c_internal_temp_calibration_t),
                 ("ntc_temp_cal_offset", c_temp_ofs_t),
-                ("report_metrics_max_interval", c_uint8),
-                ("halfwave_adjust_ticks", c_int8),
-                ("switch_on_minimum_ticks", c_uint16),
-                ("switch_on_count", c_uint8)]
+                ("report_metrics_max_interval", c_uint8)]
 
     def __dir__(self):
         parts = []
@@ -210,10 +207,7 @@ class dimmer_eeprom_written_t(Structure):
 
 class dimmer_sync_event_t(Structure):
     _pack_ = 1
-    _fields_ = [("lost", c_uint8, 1),
-                ("sync", c_uint8, 1),
-                ("__reserved", c_uint8, 2),
-                ("sync_difference_cycles", c_uint32, 28),
+    _fields_ = [("invalid signals", c_uint16),
                 ("halfwave_counter", c_uint16)]
 
     def __str__(self):